Add additional business logic

+++ C6/C6/Services/AppointmentService.cs
@@ -16,12 +16,14 @@ public class AppointmentService : IAppointmentService
                                 "Brak ConnectionString 'DefaultConnection' w konfiguracji!");
     }
 
-    public async Task<IEnumerable<AppointmentListDto>> GetAllAsync(string? status = null, string? patientLastName = null)
+    public async Task<IEnumerable<AppointmentListDto>> GetAllAsync(string? status = null,
+        string? patientLastName = null)
     {
         var result = new List<AppointmentListDto>();
         await using var connection = new SqlConnection(_connectionString);
 
-        var commandString = "SELECT IdAppointment, AppointmentDate, Status, Reason, FirstName + ' ' + LastName AS FullName, Email FROM Appointments a JOIN Patients p ON a.IdPatient = p.IdPatient";
+        var commandString =
+            "SELECT IdAppointment, AppointmentDate, Status, Reason, FirstName + ' ' + LastName AS FullName, Email FROM Appointments a JOIN Patients p ON a.IdPatient = p.IdPatient";
         var optionalConditions = new List<string>();
         if (status is not null)
         {
@@ -29,6 +31,7 @@ public class AppointmentService : IAppointmentService
                 throw new BadRequestException($"Invalid status: {status}.");
             optionalConditions.Add("a.Status = @Status");
         }
+
         if (patientLastName is not null)
             optionalConditions.Add("p.LastName LIKE @PatientLastName");
         if (optionalConditions.Count > 0)
@@ -67,6 +70,11 @@ public class AppointmentService : IAppointmentService
         await CheckDoctorExistsAsync(connection, dto.IdDoctor);
         await CheckPatientExistsAsync(connection, dto.IdPatient);
 
+        if (dto.AppointmentDate < DateTime.Now) throw new ConflictException("Cannot add an appointment in the past.");
+
+        if (string.IsNullOrEmpty(dto.InternalNotes) || dto.InternalNotes.Length > 250)
+            throw new ConflictException("InternalNotes cannot be empty and must have at most 250 characters.");
+
         var checkConflict = new SqlCommand(
             "SELECT COUNT(*) FROM Appointments WHERE IdDoctor=@IdDoctor AND AppointmentDate=@AppointmentDate",
             connection);
@@ -98,6 +106,31 @@ public class AppointmentService : IAppointmentService
         if (!Enum.IsDefined(typeof(AppointmentStatus), dto.Status))
             throw new BadRequestException($"Invalid status: {dto.Status}.");
 
+        var checkStatusCommand = new SqlCommand(
+            "SELECT Status, AppointmentDate FROM Appointments WHERE IdAppointment=@IdAppointment",
+            connection);
+        checkStatusCommand.Parameters.AddWithValue("@IdAppointment", id);
+
+        await using var reader = await checkStatusCommand.ExecuteReaderAsync();
+        while (await reader.ReadAsync())
+            if (reader.GetDateTime(reader.GetOrdinal("AppointmentDate")).Equals(dto.AppointmentDate))
+            {
+                if (Enum.Parse<AppointmentStatus>(reader.GetString(reader.GetOrdinal("Status"))) ==
+                    AppointmentStatus.Completed)
+                    throw new ConflictException(
+                        $"Cannot change date of the appointment {id}, beacause it is already completed.");
+                var checkAvailabilityCommand = new SqlCommand(
+                    "SELECT COUNT(*) FROM Appointments WHERE IdAppointment!=@IdAppointment AND  IdDoctor=@IdDoctor AND AppointmentDate=@AppointmentDate",
+                    connection);
+                checkAvailabilityCommand.Parameters.AddWithValue("@IdAppointment", id);
+                checkAvailabilityCommand.Parameters.AddWithValue("@IdDoctor", dto.IdDoctor);
+                checkAvailabilityCommand.Parameters.AddWithValue("@AppointmentDate", dto.AppointmentDate);
+
+                var count = (int?)await checkAvailabilityCommand.ExecuteScalarAsync();
+
+                if (count > 0) throw new ConflictException($"Doctor {id} has appointments at that time.");
+            }
+
         var command = new SqlCommand(
             "UPDATE Appointments SET AppointmentDate=@AppointmentDate, Status=@Status, Reason=@Reason, IdPatient=@IdPatient, IdDoctor=@IdDoctor, InternalNotes=@InternalNotes WHERE IdAppointment=@IdAppointment",
             connection);
@@ -116,17 +149,26 @@ public class AppointmentService : IAppointmentService
     public async Task DeleteAsync(string id)
     {
         await using var connection = new SqlConnection(_connectionString);
-        var command = new SqlCommand("DELETE FROM Appointments WHERE IdAppointment=@IdAppointment", connection);
-        command.Parameters.AddWithValue("@IdAppointment", id);
         await connection.OpenAsync();
-        var rows = await command.ExecuteNonQueryAsync();
-        if (rows == 0)
+
+        var selectCommand = new SqlCommand("SELECT Status FROM Appointments WHERE IdAppointment=@IdAppointment",
+            connection);
+        selectCommand.Parameters.AddWithValue("@IdAppointment", id);
+        var status = (string?)await selectCommand.ExecuteScalarAsync();
+        if (status is null)
             throw new NotFoundException($"{id} not found.");
+        if (Enum.Parse<AppointmentStatus>(status) == AppointmentStatus.Completed)
+            throw new ConflictException($"{id} is already completed.");
+
+        var deleteCommand = new SqlCommand("DELETE FROM Appointments WHERE IdAppointment=@IdAppointment", connection);
+        deleteCommand.Parameters.AddWithValue("@IdAppointment", id);
+        await deleteCommand.ExecuteNonQueryAsync();
     }
 
     private static async Task CheckDoctorExistsAsync(SqlConnection connection, int idDoctor)
     {
-        var command = new SqlCommand("SELECT COUNT(*) FROM Doctors WHERE IdDoctor=@IdDoctor", connection);
+        var command = new SqlCommand("SELECT COUNT(*) FROM Doctors WHERE IdDoctor=@IdDoctor AND IsActive=1",
+            connection);
         command.Parameters.AddWithValue("@IdDoctor", idDoctor);
         if ((int)await command.ExecuteScalarAsync()! == 0)
             throw new NotFoundException($"Doctor {idDoctor} not found.");
@@ -134,7 +176,8 @@ public class AppointmentService : IAppointmentService
 
     private static async Task CheckPatientExistsAsync(SqlConnection connection, int idPatient)
     {
-        var command = new SqlCommand("SELECT COUNT(*) FROM Patients WHERE IdPatient=@IdPatient", connection);
+        var command = new SqlCommand("SELECT COUNT(*) FROM Patients WHERE IdPatient=@IdPatient AND IsActive=1",
+            connection);
         command.Parameters.AddWithValue("@IdPatient", idPatient);
         if ((int)await command.ExecuteScalarAsync()! == 0)
             throw new NotFoundException($"Patient {idPatient} not found.");
